refactor(ui): split ObservabilityView into per-component files
Pure mechanical refactor — no behavior change. The 1214-line monolith
is now ~500 lines of orchestration plus six focused files under
`ui/src/features/overview/observability/`:

  styles.ts           shared visual tokens (PROVIDER_COLORS,
                      providerColor, phaseColor, PHASE_LABEL,
                      ATTR_PRIORITY_KEYS, thStyle, tdBase,
                      DRAWER_BREAKPOINT_PX, StatusFilter)
  StatCard.tsx        single-cell stat used in the runtime strip
  CopyableID.tsx      truncated request-id + copy button
  SpanWaterfall.tsx   waterfall + sticky ruler + SpanRow +
                      SpanAttributePanel (the largest chunk by
                      line count, but a single coherent unit)
  RouteCandidates.tsx per-trace router decision strip
  TraceDetail.tsx     drawer body composition (status grid +
                      waterfall + route summary + candidates +
                      event flow + diagnostics)

ObservabilityView.tsx stays at the same path so the lazy import
in AppShell.tsx and the test in ObservabilityView.test.tsx don't
need to change.

Verified:
  - bun run build: clean. ObservabilityView chunk 31.25 KB
    (was 31.22 KB before the split — module-boundary overhead
    only, no real growth).
  - bun run test: 30 files, 665 tests passing (unchanged).
  - bun run test:e2e: 52 passing (unchanged).

This unblocks the Tier 2 redesign work tracked in
docs/rfcs/observability-redesign.md (PR #72) — once the
two-tab IA, filter bar, drill-down, and phase-rollup work
starts, the per-component split makes the diffs reviewable.
