+++ ui/src/features/overview/ObservabilityView.tsx
@@ -1,25 +1,42 @@
-import { useCallback, useEffect, useMemo, useRef, useState, type CSSProperties } from "react";
+// ObservabilityView is the top-level shell for the Observability
+// workspace: header + filter chips, runtime stat strip, recent-traces
+// table, and the inline drawer (or modal at narrow widths) wired to
+// the trace detail. Per-section components live under
+// `./observability/`; this file is orchestration only — state, polling
+// effects, filter computation, and layout.
+
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+
+import type { RuntimeConsoleViewModel } from "../../app/useRuntimeConsole";
 import { getMCPCacheStats, getRecentTraces, getRuntimeStats, getTrace } from "../../lib/api";
 import {
   buildSpanWaterfall,
   buildTraceTimeline,
-  describeHealthStatus,
-  describeRouteCandidateOutcome,
-  describeRouteReason,
-  describeRouteSkipReason,
-  explainRouteCandidate,
-  formatRelativeTime,
-  formatTraceAttributeKey,
-  formatTraceAttributeValue,
-  healthStatusTone,
-  routeOutcomeTone,
-  traceStatusBadge,
   type TraceTimelineItem,
-  type WaterfallSpan,
+} from "../../lib/runtime-trace";
+import {
+  describeRouteReason,
+  formatRelativeTime,
+  traceStatusBadge,
 } from "../../lib/runtime-utils";
-import type { RuntimeConsoleViewModel } from "../../app/useRuntimeConsole";
-import type { MCPCacheStatsResponse, RuntimeStatsResponse, TraceListItem, TraceResponse, TraceSpanRecord } from "../../types/runtime";
+import type {
+  MCPCacheStatsResponse,
+  RuntimeStatsResponse,
+  TraceListItem,
+  TraceResponse,
+} from "../../types/runtime";
 import { Badge, Icon, Icons, Modal, ModelPicker, ProviderPicker, Toggle } from "../shared/ui";
+
+import { CopyableID } from "./observability/CopyableID";
+import { StatCard } from "./observability/StatCard";
+import { TraceDetail } from "./observability/TraceDetail";
+import {
+  DRAWER_BREAKPOINT_PX,
+  providerColor,
+  type StatusFilter,
+  tdBase,
+  thStyle,
+} from "./observability/styles";
 
 type Props = {
   state: RuntimeConsoleViewModel["state"];
@@ -30,131 +47,6 @@ type Props = {
   onNavigate?: (workspace: "chats" | "providers" | "runs" | "overview" | "costs" | "settings") => void;
   focusRequest?: { requestID: string; nonce: number } | null;
 };
-
-// 900px chosen because at narrower widths the inline split between
-// table and drawer feels too cramped — the table only has so many
-// columns it can shed before unreadable. Below that we fall back to
-// the centered Modal we used to ship.
-const DRAWER_BREAKPOINT_PX = 900;
-
-const PROVIDER_COLORS: Record<string, string> = {
-  anthropic:   "var(--brand-anthropic)",
-  openai:      "var(--brand-openai)",
-  gemini:      "var(--brand-gemini)",
-  mistral:     "var(--brand-mistral)",
-  groq:        "var(--brand-groq)",
-  deepseek:    "var(--teal)",
-  perplexity:  "var(--teal)",
-  together_ai: "var(--t2)",
-  xai:         "var(--t0)",
-  ollama:      "var(--teal)",
-  lmstudio:    "var(--t2)",
-  llamacpp:    "var(--t2)",
-  localai:     "var(--t2)",
-};
-
-function providerColor(id: string): string {
-  return PROVIDER_COLORS[id.toLowerCase()] ?? "var(--teal)";
-}
-
-// phaseColor maps the phase classification to an existing token. The
-// provider phase additionally pulls the brand color from the span's
-// `gen_ai.provider.name` attribute when present so the bar visually
-// matches the row in the table.
-function phaseColor(phase: TraceTimelineItem["phase"], span?: TraceSpanRecord): string {
-  if (phase === "provider") {
-    const attr = span?.attributes?.["gen_ai.provider.name"];
-    if (typeof attr === "string" && attr) {
-      const c = PROVIDER_COLORS[attr.toLowerCase()];
-      if (c) return c;
-    }
-    return "var(--brand-anthropic)";
-  }
-  switch (phase) {
-    case "request":  return "var(--teal)";
-    case "routing":  return "var(--amber)";
-    case "governor": return "var(--brand-mistral)";
-    case "cost":     return "var(--t2)";
-    case "usage":    return "var(--t2)";
-    case "queue":    return "var(--amber)";
-    case "orchestration": return "var(--t2)";
-    case "tool":     return "var(--brand-openai)";
-    case "approval": return "var(--brand-anthropic)";
-    case "artifact": return "var(--green)";
-    case "retention": return "var(--t3)";
-    case "agent_chat": return "var(--brand-openai)";
-    case "response": return "var(--teal)";
-    default:         return "var(--t3)";
-  }
-}
-
-const PHASE_LABEL: Record<TraceTimelineItem["phase"], string> = {
-  request: "request", routing: "routing",
-  provider: "provider", governor: "governor", usage: "usage",
-  cost: "cost", response: "response", queue: "queue",
-  orchestration: "orchestration", tool: "tool", approval: "approval",
-  artifact: "artifact", retention: "retention", agent_chat: "agent chat",
-  other: "other",
-};
-
-type StatCardProps = { label: string; value: string | number; sub?: string; highlight?: boolean };
-function StatCard({ label, value, sub, highlight }: StatCardProps) {
-  return (
-    <div className="card" style={{ padding: "12px 14px", minWidth: 110 }}>
-      <div className="kicker" style={{ color: "var(--t2)", marginBottom: 6 }}>{label}</div>
-      <div style={{ fontSize: 22, fontWeight: 600, fontFamily: "var(--font-mono)", color: highlight ? "var(--amber)" : "var(--t0)", lineHeight: 1 }}>{value}</div>
-      {sub && <div style={{ fontSize: 10, color: "var(--t3)", fontFamily: "var(--font-mono)", marginTop: 4 }}>{sub}</div>}
-    </div>
-  );
-}
-
-// thStyle mirrors the Providers table header — uppercase 11px t2 with
-// the same horizontal padding so the two views share visual chrome.
-const thStyle: CSSProperties = {
-  padding: "6px 12px",
-  textAlign: "left",
-  fontSize: 11,
-  fontWeight: 500,
-  color: "var(--t2)",
-  letterSpacing: "0.04em",
-  textTransform: "uppercase",
-  whiteSpace: "nowrap",
-};
-
-const tdBase: CSSProperties = {
-  padding: "8px 12px",
-  fontSize: 12,
-  fontFamily: "var(--font-mono)",
-  whiteSpace: "nowrap",
-  overflow: "hidden",
-  textOverflow: "ellipsis",
-};
-
-type StatusFilter = "all" | "healthy" | "error";
-
-function CopyableID({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false);
-  return (
-    <button
-      onClick={e => {
-        e.stopPropagation();
-        navigator.clipboard?.writeText(text).catch(() => {});
-        setCopied(true);
-        window.setTimeout(() => setCopied(false), 1500);
-      }}
-      title={text}
-      style={{
-        background: "none", border: "none", padding: 0, cursor: "pointer",
-        fontFamily: "var(--font-mono)", fontSize: 11,
-        color: copied ? "var(--green)" : "var(--teal)",
-        display: "inline-flex", alignItems: "center", gap: 4,
-        overflow: "hidden", textOverflow: "ellipsis", maxWidth: "100%",
-      }}>
-      <span style={{ overflow: "hidden", textOverflow: "ellipsis" }}>{text.slice(0, 8)}…</span>
-      <Icon d={copied ? Icons.check : Icons.copy} size={11} />
-    </button>
-  );
-}
 
 export function ObservabilityView({ state, onNavigate, focusRequest }: Props) {
   const [runtimeStats, setRuntimeStats] = useState<RuntimeStatsResponse["data"] | null>(null);
@@ -633,656 +525,6 @@ export function ObservabilityView({ state, onNavigate, focusRequest }: Props) {
           {detailContent}
         </Modal>
       )}
-    </div>
-  );
-}
-
-// ─── TraceDetail ─────────────────────────────────────────────────────────────
-
-type LedgerEntry = NonNullable<RuntimeConsoleViewModel["state"]["requestLedger"]>[number];
-
-type TraceDetailProps = {
-  selectedID: string;
-  selectedTrace?: TraceListItem;
-  ledger?: LedgerEntry;
-  traceDetail: TraceResponse["data"] | null;
-  traceFetching: boolean;
-  waterfall: ReturnType<typeof buildSpanWaterfall>;
-  traceTimeline: ReturnType<typeof buildTraceTimeline>;
-  expandedSpanID: string | null;
-  setExpandedSpanID: (id: string | null) => void;
-  phaseFilter: TraceTimelineItem["phase"] | null;
-  setPhaseFilter: (p: TraceTimelineItem["phase"] | null) => void;
-};
-
-function TraceDetail({
-  selectedTrace, ledger, traceDetail, traceFetching,
-  waterfall, traceTimeline, expandedSpanID, setExpandedSpanID,
-  phaseFilter, setPhaseFilter,
-}: TraceDetailProps) {
-  const status = selectedTrace ? traceStatusBadge(selectedTrace) : null;
-  const tokens = ledger
-    ? `${ledger.prompt_tokens ?? 0} / ${ledger.completion_tokens ?? 0}`
-    : "—";
-  const cost = ledger?.amount_usd
-    ? `$${Number.parseFloat(ledger.amount_usd).toFixed(5)}`
-    : ledger ? "$0.00000" : "—";
-  const latency = selectedTrace?.duration_ms != null ? `${selectedTrace.duration_ms}ms` : "—";
-
-  return (
-    <div style={{ display: "flex", flexDirection: "column", gap: 14 }}>
-      {/* Live status grid */}
-      <div style={{
-        display: "grid", gridTemplateColumns: "repeat(4, 1fr)", gap: 12,
-        padding: 12, border: "1px solid var(--border)", borderRadius: "var(--radius-sm)",
-      }}>
-        <div>
-          <div className="kicker" style={{ marginBottom: 4 }}>Status</div>
-          {status
-            ? <Badge status={status.status} label={status.label} />
-            : <span style={{ fontSize: 12, color: "var(--t3)" }}>—</span>}
-        </div>
-        <div>
-          <div className="kicker" style={{ marginBottom: 4 }}>Latency</div>
-          <div style={{ fontSize: 12, color: "var(--t0)", fontFamily: "var(--font-mono)" }}>{latency}</div>
-        </div>
-        <div>
-          <div className="kicker" style={{ marginBottom: 4 }}>Tokens</div>
-          <div style={{ fontSize: 12, color: "var(--t0)", fontFamily: "var(--font-mono)" }}>{tokens}</div>
-        </div>
-        <div>
-          <div className="kicker" style={{ marginBottom: 4 }}>Cost</div>
-          <div style={{ fontSize: 12, color: "var(--t0)", fontFamily: "var(--font-mono)" }}>{cost}</div>
-        </div>
-      </div>
-
-      {/* Span waterfall — the centerpiece. */}
-      <SpanWaterfall
-        waterfall={waterfall}
-        expandedSpanID={expandedSpanID}
-        setExpandedSpanID={setExpandedSpanID}
-        phaseFilter={phaseFilter}
-        setPhaseFilter={setPhaseFilter}
-        traceFetching={traceFetching}
-        hasTraceDetail={!!traceDetail}
-      />
-
-      {/* Route summary */}
-      {selectedTrace?.route && (
-        <div style={{ padding: "10px 12px", border: "1px solid var(--border)", borderRadius: "var(--radius-sm)" }}>
-          <div className="kicker" style={{ marginBottom: 8 }}>Route summary</div>
-          <div style={{ display: "flex", flexWrap: "wrap", gap: 12, alignItems: "center" }}>
-            {selectedTrace.route.final_provider && (
-              <span style={{ fontSize: 11, fontFamily: "var(--font-mono)", color: "var(--t1)" }}>
-                <span style={{ color: "var(--t3)" }}>provider </span>{selectedTrace.route.final_provider}
-              </span>
-            )}
-            {selectedTrace.route.final_model && (
-              <span style={{ fontSize: 11, fontFamily: "var(--font-mono)", color: "var(--t1)" }}>
-                <span style={{ color: "var(--t3)" }}>model </span>{selectedTrace.route.final_model}
-              </span>
-            )}
-            {selectedTrace.route.final_reason && (
-              <Badge status="queued" label={describeRouteReason(selectedTrace.route.final_reason)} />
-            )}
-            {selectedTrace.route.fallback_from && (
-              <span style={{ fontSize: 11, color: "var(--amber)", fontFamily: "var(--font-mono)" }}>
-                ↳ from {selectedTrace.route.fallback_from}
-              </span>
-            )}
-          </div>
-        </div>
-      )}
-
-      {/* Route candidates */}
-      <RouteCandidates traceDetail={traceDetail} selectedTrace={selectedTrace} />
-
-      {/* Event flow */}
-      {traceTimeline.length > 0 && (
-        <div style={{ padding: "10px 12px", border: "1px solid var(--border)", borderRadius: "var(--radius-sm)" }}>
-          <div className="kicker" style={{ marginBottom: 8 }}>Event flow</div>
-          <div style={{ display: "flex", flexDirection: "column", gap: 8 }}>
-            {traceTimeline.map((event, index) => (
-              <div key={`${event.timestamp}-${event.name}-${index}`} style={{ display: "grid", gridTemplateColumns: "56px 92px 1fr", gap: 10, alignItems: "start" }}>
-                <div style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>{event.offsetLabel}</div>
-                <Badge
-                  status={
-                    event.phase === "provider" ? "healthy" :
-                    event.phase === "queue" ? "queued" :
-                    event.phase === "routing" ? "ok" :
-                    event.phase === "response" ? "done" :
-                    "disabled"
-                  }
-                  label={event.phase}
-                />
-                <div>
-                  <div style={{ fontSize: 12, color: "var(--t0)", marginBottom: 4 }}>{event.name}</div>
-                  {event.name === "governor.model_rewrite" ? (
-                    <div style={{ display: "flex", flexDirection: "column", gap: 4 }}>
-                      {event.attributes?.["gen_ai.request.model.original"] != null && event.attributes?.["gen_ai.request.model.rewritten"] != null && (
-                        <span style={{ fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t1)" }}>
-                          {String(event.attributes["gen_ai.request.model.original"])} → {String(event.attributes["gen_ai.request.model.rewritten"])}
-                        </span>
-                      )}
-                      <div style={{ display: "flex", flexWrap: "wrap", gap: 8 }}>
-                        {event.attributes?.["hecate.policy.rule_id"] != null && (
-                          <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                            rule <span style={{ color: "var(--t1)" }}>{String(event.attributes["hecate.policy.rule_id"])}</span>
-                          </span>
-                        )}
-                        {event.attributes?.["hecate.policy.action"] != null && (
-                          <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                            action <span style={{ color: "var(--t1)" }}>{String(event.attributes["hecate.policy.action"])}</span>
-                          </span>
-                        )}
-                        {event.attributes?.["hecate.policy.reason"] != null && (
-                          <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t2)", fontStyle: "italic" }}>
-                            {String(event.attributes["hecate.policy.reason"])}
-                          </span>
-                        )}
-                      </div>
-                    </div>
-                  ) : event.attributes && Object.keys(event.attributes).length > 0 && (
-                    <div style={{ display: "flex", flexWrap: "wrap", gap: 8 }}>
-                      {Object.entries(event.attributes).slice(0, 4).map(([key, value]) => (
-                        <span key={key} style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                          {formatTraceAttributeKey(key)} <span style={{ color: "var(--t1)" }}>{formatTraceAttributeValue(value)}</span>
-                        </span>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* Diagnostics — collapsed raw span dump for the expanded span. */}
-      {expandedSpanID && (() => {
-        const ws = waterfall.spans.find(s => s.span.span_id === expandedSpanID);
-        if (!ws) return null;
-        const attrs = ws.span.attributes ?? {};
-        const attrEntries = Object.entries(attrs).filter(([, v]) => v != null && v !== "");
-        return (
-          <details>
-            <summary style={{ fontSize: 11, color: "var(--t2)", cursor: "pointer", userSelect: "none" }}>
-              Diagnostics
-            </summary>
-            <div style={{ marginTop: 8, padding: 10, background: "var(--bg2)", borderRadius: "var(--radius-sm)" }}>
-              <div className="kicker" style={{ marginBottom: 6 }}>{ws.span.name}</div>
-              <div style={{ display: "flex", flexDirection: "column", gap: 3 }}>
-                {[
-                  ["span_id",  ws.span.span_id],
-                  ["start",    `+${ws.startMs}ms`],
-                  ["duration", `${ws.durMs}ms`],
-                  ["status",   ws.span.status_code],
-                ].filter(([,v]) => v).map(([k, v]) => (
-                  <div key={k} style={{ display: "flex", gap: 8 }}>
-                    <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)", width: 60, flexShrink: 0 }}>{k}</span>
-                    <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--teal)" }}>{v}</span>
-                  </div>
-                ))}
-                {attrEntries.length > 0 && (
-                  <div style={{ marginTop: 4, borderTop: "1px solid var(--border)", paddingTop: 4, display: "flex", flexDirection: "column", gap: 2 }}>
-                    {attrEntries.map(([k, v]) => (
-                      <div key={k} style={{ display: "flex", gap: 8 }}>
-                        <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)", width: 60, flexShrink: 0, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }} title={k}>{k.split(".").pop()}</span>
-                        <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t1)", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>{String(v)}</span>
-                      </div>
-                    ))}
-                  </div>
-                )}
-              </div>
-            </div>
-          </details>
-        );
-      })()}
-
-      {!traceFetching && !traceDetail && (
-        <div style={{ fontSize: 12, color: "var(--t3)" }}>No trace detail available.</div>
-      )}
-    </div>
-  );
-}
-
-// ─── SpanWaterfall ───────────────────────────────────────────────────────────
-
-const ATTR_PRIORITY_KEYS = [
-  "provider", "gen_ai.provider.name",
-  "model", "gen_ai.request.model", "gen_ai.response.model",
-  "status_code", "error",
-  "usage.input_tokens", "gen_ai.usage.input_tokens",
-  "usage.output_tokens", "gen_ai.usage.output_tokens",
-  "route.skip_reason", "route.fallback_from",
-];
-
-type SpanWaterfallProps = {
-  waterfall: ReturnType<typeof buildSpanWaterfall>;
-  expandedSpanID: string | null;
-  setExpandedSpanID: (id: string | null) => void;
-  phaseFilter: TraceTimelineItem["phase"] | null;
-  setPhaseFilter: (p: TraceTimelineItem["phase"] | null) => void;
-  traceFetching: boolean;
-  hasTraceDetail: boolean;
-};
-
-function SpanWaterfall({
-  waterfall, expandedSpanID, setExpandedSpanID,
-  phaseFilter, setPhaseFilter, traceFetching, hasTraceDetail,
-}: SpanWaterfallProps) {
-  const { spans, totalMs, phases } = waterfall;
-
-  if (spans.length === 0) {
-    return (
-      <div style={{ padding: "10px 12px", border: "1px solid var(--border)", borderRadius: "var(--radius-sm)" }}>
-        <div className="kicker-lg" style={{ marginBottom: 8, fontSize: 12, fontWeight: 500, color: "var(--t1)" }}>Spans</div>
-        <div style={{ fontSize: 12, color: "var(--t3)", fontFamily: "var(--font-mono)" }}>
-          {traceFetching && !hasTraceDetail ? "loading…" : "Span data not available for this trace."}
-        </div>
-      </div>
-    );
-  }
-
-  const ticks = [0, Math.round(totalMs / 3), Math.round((2 * totalMs) / 3), totalMs];
-
-  return (
-    <div style={{ padding: "10px 12px", border: "1px solid var(--border)", borderRadius: "var(--radius-sm)" }}>
-      {/* Header — count + total */}
-      <div style={{ display: "flex", alignItems: "center", gap: 10, marginBottom: 8, flexWrap: "wrap" }}>
-        <span className="kicker-lg" style={{ fontSize: 12, fontWeight: 500, color: "var(--t0)", letterSpacing: "0.04em", textTransform: "uppercase" }}>
-          Spans ({spans.length}) · total {totalMs} ms
-        </span>
-        {phases.length > 1 && (
-          <div role="group" aria-label="Phase legend" style={{ display: "flex", gap: 4, flexWrap: "wrap", marginLeft: "auto" }}>
-            {phases.map(p => {
-              const active = phaseFilter === p;
-              return (
-                <button
-                  key={p}
-                  onClick={() => setPhaseFilter(active ? null : p)}
-                  aria-pressed={active}
-                  style={{
-                    background: active ? "var(--bg3)" : "transparent",
-                    border: `1px solid ${active ? "var(--teal)" : "var(--border)"}`,
-                    borderRadius: "var(--radius-sm)",
-                    padding: "2px 6px",
-                    cursor: "pointer",
-                    display: "inline-flex",
-                    alignItems: "center",
-                    gap: 4,
-                    fontFamily: "var(--font-mono)",
-                    fontSize: 10,
-                    color: "var(--t1)",
-                  }}>
-                  <span style={{ display: "inline-block", width: 8, height: 8, borderRadius: 2, background: phaseColor(p) }} />
-                  {PHASE_LABEL[p]}
-                </button>
-              );
-            })}
-          </div>
-        )}
-      </div>
-
-      {/* Sticky ruler. Three things needed for it to render reliably
-        on top of subsequent span rows:
-         - `bg2` (not `bg1`) so the ruler has visible contrast against
-           the surrounding card and so its background actually
-           occludes the row that scrolls under it. `bg1` is the same
-           color as the card it's sitting in.
-         - `isolation: isolate` creates a stacking context so the
-           absolute-positioned bar children inside SpanRow can't
-           paint above the sticky ruler. Without it, a high-z-index
-           descendant of a sibling could land on top.
-         - `marginBottom: 6` separates the ruler from the first span
-           row — without it, the row's amber critical-path border
-           visually merges with the ruler's own bottom border. */}
-      <div style={{
-        position: "sticky", top: 0, zIndex: 5,
-        background: "var(--bg2)",
-        isolation: "isolate",
-        display: "grid",
-        gridTemplateColumns: "240px 1fr 60px",
-        gap: 8,
-        padding: "6px 0",
-        marginBottom: 6,
-        borderBottom: "1px solid var(--border)",
-      }}>
-        <div />
-        <div style={{ position: "relative", height: 12 }}>
-          {ticks.map((t, i) => (
-            <span
-              key={i}
-              style={{
-                position: "absolute",
-                left: `${(t / totalMs) * 100}%`,
-                transform: i === 0 ? "translateX(0)" : i === ticks.length - 1 ? "translateX(-100%)" : "translateX(-50%)",
-                fontFamily: "var(--font-mono)",
-                fontSize: 10,
-                color: "var(--t2)",
-              }}>{t}ms</span>
-          ))}
-        </div>
-        <div />
-      </div>
-
-      {/* Span rows */}
-      <div style={{ display: "flex", flexDirection: "column" }}>
-        {spans.map((ws) => (
-          <SpanRow
-            key={ws.span.span_id}
-            ws={ws}
-            totalMs={totalMs}
-            isExpanded={expandedSpanID === ws.span.span_id}
-            isDimmed={phaseFilter !== null && ws.phase !== phaseFilter}
-            onToggle={() => setExpandedSpanID(expandedSpanID === ws.span.span_id ? null : ws.span.span_id)}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function SpanRow({
-  ws, totalMs, isExpanded, isDimmed, onToggle,
-}: {
-  ws: WaterfallSpan;
-  totalMs: number;
-  isExpanded: boolean;
-  isDimmed: boolean;
-  onToggle: () => void;
-}) {
-  // unknownTiming: span had missing/unparseable timestamps. We render
-  // a tiny 1%-wide indicator at left:0 (not a meaningful position —
-  // see the comment in runtime-trace.ts) plus a "?" duration label,
-  // so the row is visible without silently misleading the operator
-  // about position or width.
-  // negativeDuration: both timestamps parsed but end < start (clock
-  // skew or partial trace). The bar IS positioned at the span's
-  // real start but the duration label is "?" rather than a
-  // misleading "-50ms" or a silently-clamped 1ms tick.
-  const leftPct = ws.unknownTiming ? 0 : (ws.startMs / totalMs) * 100;
-  const widthPct = ws.unknownTiming || ws.negativeDuration
-    ? 1
-    : Math.max((ws.durMs / totalMs) * 100, 0.5);
-  const color = phaseColor(ws.phase, ws.span);
-  const opacity = isDimmed ? 0.3 : 1;
-  // Duration label inside the bar when wide enough, otherwise to its right.
-  const labelInside = widthPct > 12 && !ws.unknownTiming && !ws.negativeDuration;
-  const durLabel = ws.unknownTiming || ws.negativeDuration ? "?" : `${Math.max(ws.durMs, 0).toFixed(0)}ms`;
-  const barTone = ws.unknownTiming || ws.negativeDuration ? "var(--t3)" : (ws.hasError ? "var(--red)" : color);
-
-  return (
-    <div>
-      <div
-        onClick={onToggle}
-        role="button"
-        tabIndex={0}
-        aria-label={`span ${ws.span.name}`}
-        aria-expanded={isExpanded}
-        onKeyDown={e => {
-          if (e.key === "Enter" || e.key === " ") { e.preventDefault(); onToggle(); }
-        }}
-        style={{
-          height: 22,
-          display: "grid",
-          gridTemplateColumns: "240px 1fr 60px",
-          gap: 8,
-          alignItems: "center",
-          cursor: "pointer",
-          background: isExpanded ? "var(--bg2)" : "transparent",
-          borderLeft: ws.critical ? "2px solid var(--amber)" : "2px solid transparent",
-          opacity,
-          paddingRight: 4,
-        }}>
-        {/* Span name column with depth indent + status dot */}
-        <div style={{ display: "flex", alignItems: "center", gap: 6, paddingLeft: 6 + ws.depth * 12, overflow: "hidden" }}>
-          <span
-            aria-label={ws.hasError ? "error" : "ok"}
-            style={{
-              flexShrink: 0,
-              width: 6, height: 6, borderRadius: "50%",
-              background: ws.hasError ? "var(--red)" : "var(--green)",
-            }}
-          />
-          <span
-            title={ws.span.name}
-            style={{
-              fontFamily: "var(--font-mono)",
-              fontSize: 12,
-              color: ws.hasError ? "var(--red)" : "var(--t1)",
-              overflow: "hidden",
-              textOverflow: "ellipsis",
-              whiteSpace: "nowrap",
-              flex: 1,
-            }}>
-            {ws.span.name}
-          </span>
-          {ws.critical && (
-            <span title="critical path" style={{ color: "var(--amber)", fontSize: 10, flexShrink: 0 }}>★</span>
-          )}
-        </div>
-
-        {/* Bar column */}
-        <div style={{ position: "relative", height: 14 }}>
-          <div
-            style={{
-              position: "absolute",
-              left: `${leftPct}%`,
-              width: `${widthPct}%`,
-              minWidth: 2,
-              height: "100%",
-              background: barTone,
-              borderRadius: 2,
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "flex-end",
-              paddingRight: 4,
-              boxSizing: "border-box",
-            }}>
-            {labelInside && (
-              <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t0)" }}>
-                {durLabel}
-              </span>
-            )}
-          </div>
-        </div>
-
-        {/* Right-side label */}
-        <div
-          title={ws.unknownTiming
-            ? "missing or unparseable timestamps"
-            : ws.negativeDuration
-              ? "end_time is before start_time (clock skew or partial trace)"
-              : undefined}
-          style={{
-            fontFamily: "var(--font-mono)",
-            fontSize: 10,
-            color: ws.unknownTiming || ws.negativeDuration ? "var(--amber)" : "var(--t1)",
-            textAlign: "right",
-            whiteSpace: "nowrap",
-          }}>
-          {labelInside ? "" : durLabel}
-          {ws.hasError ? " · ERR" : ""}
-        </div>
-      </div>
-
-      {isExpanded && <SpanAttributePanel ws={ws} />}
-    </div>
-  );
-}
-
-function SpanAttributePanel({ ws }: { ws: WaterfallSpan }) {
-  const attrs = ws.span.attributes ?? {};
-  const present = Object.entries(attrs).filter(([, v]) => v != null && v !== "");
-  const priority = present.filter(([k]) => ATTR_PRIORITY_KEYS.includes(k));
-  const rest = present.filter(([k]) => !ATTR_PRIORITY_KEYS.includes(k));
-
-  return (
-    <div
-      data-testid={`span-attrs-${ws.span.span_id}`}
-      style={{
-        marginLeft: 12 + ws.depth * 12,
-        marginTop: 2,
-        marginBottom: 6,
-        padding: "8px 10px",
-        background: "var(--bg2)",
-        border: "1px solid var(--border)",
-        borderRadius: "var(--radius-sm)",
-      }}>
-      {priority.length === 0 && rest.length === 0 ? (
-        <div style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-          No attributes recorded for this span.
-        </div>
-      ) : (
-        <>
-          <div style={{ display: "flex", flexDirection: "column", gap: 2 }}>
-            {priority.map(([k, v]) => (
-              <div key={k} style={{ display: "flex", gap: 8 }}>
-                <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)", width: 140, flexShrink: 0 }}>{k}</span>
-                <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t1)", overflow: "hidden", textOverflow: "ellipsis" }}>{String(v)}</span>
-              </div>
-            ))}
-          </div>
-          {rest.length > 0 && (
-            <details style={{ marginTop: priority.length ? 6 : 0 }}>
-              <summary style={{ fontSize: 10, color: "var(--t2)", cursor: "pointer", userSelect: "none", fontFamily: "var(--font-mono)" }}>
-                {rest.length} more attribute{rest.length === 1 ? "" : "s"}
-              </summary>
-              <div style={{ marginTop: 4, display: "flex", flexDirection: "column", gap: 2 }}>
-                {rest.map(([k, v]) => (
-                  <div key={k} style={{ display: "flex", gap: 8 }}>
-                    <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)", width: 140, flexShrink: 0, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }} title={k}>{k}</span>
-                    <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t1)", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>{String(v)}</span>
-                  </div>
-                ))}
-              </div>
-            </details>
-          )}
-        </>
-      )}
-    </div>
-  );
-}
-
-// ─── RouteCandidates ─────────────────────────────────────────────────────────
-
-function RouteCandidates({
-  traceDetail, selectedTrace,
-}: {
-  traceDetail: TraceResponse["data"] | null;
-  selectedTrace?: TraceListItem;
-}) {
-  type Candidate = NonNullable<NonNullable<TraceListItem["route"]>["candidates"]>[number];
-  const candidates: Candidate[] = traceDetail?.route?.candidates ?? selectedTrace?.route?.candidates ?? [];
-  if (candidates.length === 0) return null;
-  const selected = candidates.find((c) => c.outcome === "selected" || c.outcome === "completed");
-  const skipped = candidates.filter((c) => c.outcome === "skipped").length;
-  const failed = candidates.filter((c) => c.outcome === "failed").length;
-  return (
-    <div style={{ padding: "10px 12px", border: "1px solid var(--border)", borderRadius: "var(--radius-sm)" }}>
-      <div className="kicker" style={{ marginBottom: 6 }}>Route decision</div>
-      <div style={{ marginBottom: 8, display: "flex", flexWrap: "wrap", gap: 8, alignItems: "center" }}>
-        <span style={{ fontSize: 12, color: "var(--t1)" }}>
-          {selected?.provider
-            ? `Selected ${selected.provider}/${selected.model || "provider default"}`
-            : "No provider selected"}
-        </span>
-        {skipped > 0 && <Badge status="warn" label={`${skipped} skipped`} />}
-        {failed > 0 && <Badge status="error" label={`${failed} failed`} />}
-        <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-          {candidates.length} candidate{candidates.length === 1 ? "" : "s"}
-        </span>
-      </div>
-      {candidates.map((c, i) => {
-        const outcomeTone = routeOutcomeTone(c.outcome);
-        const outcomeStatus =
-          outcomeTone === "healthy" ? "done" :
-          outcomeTone === "danger" ? "error" :
-          outcomeTone === "warning" ? "warn" :
-          "disabled";
-        const healthTone = healthStatusTone(c.health_status);
-        const healthStatus =
-          healthTone === "healthy" ? "healthy" :
-          healthTone === "danger" ? "error" :
-          healthTone === "warning" ? "warn" :
-          "disabled";
-        return (
-          <div key={`${c.provider || "provider"}-${c.model || "model"}-${c.outcome || "candidate"}-${c.index ?? i}`} style={{ padding: "8px 0", borderBottom: i === candidates.length - 1 ? undefined : "1px solid var(--border)" }}>
-            <div style={{ display: "flex", alignItems: "center", gap: 8, marginBottom: 6 }}>
-              <span style={{ fontFamily: "var(--font-mono)", fontSize: 11, color: "var(--t0)", flex: 1, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-                {c.provider}/{c.model || "no model"}
-              </span>
-              <Badge status={outcomeStatus} label={describeRouteCandidateOutcome(c)} />
-            </div>
-            <div style={{ marginBottom: 6, fontSize: 11, color: "var(--t2)", lineHeight: 1.45 }}>
-              {explainRouteCandidate(c)}
-            </div>
-            <div style={{ display: "flex", flexWrap: "wrap", gap: 6 }}>
-              {c.skip_reason && <Badge status="warn" label={describeRouteSkipReason(c.skip_reason)} />}
-              {c.health_status && <Badge status={healthStatus} label={describeHealthStatus(c.health_status)} />}
-              {c.reason && (
-                <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                  reason <span style={{ color: "var(--t1)" }}>{describeRouteReason(c.reason)}</span>
-                </span>
-              )}
-              {c.latency_ms != null && c.latency_ms > 0 && (
-                <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                  latency <span style={{ color: "var(--t1)" }}>{c.latency_ms}ms</span>
-                </span>
-              )}
-              {c.estimated_usd && (
-                <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                  est <span style={{ color: "var(--t1)" }}>{c.estimated_usd}</span>
-                </span>
-              )}
-            </div>
-            {c.detail && (
-              <div style={{ marginTop: 6, fontSize: 11, color: "var(--t2)", lineHeight: 1.45 }}>
-                {c.detail}
-              </div>
-            )}
-            {(c.policy_rule_id || c.policy_action || c.policy_reason) && (
-              <div style={{ marginTop: 6, display: "flex", flexWrap: "wrap", gap: 10 }}>
-                {c.policy_rule_id && (
-                  <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                    rule <span style={{ color: "var(--t1)" }}>{c.policy_rule_id}</span>
-                  </span>
-                )}
-                {c.policy_action && (
-                  <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                    action <span style={{ color: "var(--t1)" }}>{c.policy_action}</span>
-                  </span>
-                )}
-                {c.policy_reason && (
-                  <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t2)", fontStyle: "italic" }}>
-                    {c.policy_reason}
-                  </span>
-                )}
-              </div>
-            )}
-            {(c.failover_from || c.failover_to || c.attempt || c.retry_count) && (
-              <div style={{ marginTop: 6, display: "flex", flexWrap: "wrap", gap: 10 }}>
-                {c.attempt != null && (
-                  <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                    attempt <span style={{ color: "var(--t1)" }}>{c.attempt}</span>
-                  </span>
-                )}
-                {c.retry_count != null && (
-                  <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                    retries <span style={{ color: "var(--t1)" }}>{c.retry_count}</span>
-                  </span>
-                )}
-                {c.failover_from && c.failover_to && (
-                  <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--t3)" }}>
-                    failover <span style={{ color: "var(--t1)" }}>{c.failover_from}</span> → <span style={{ color: "var(--t1)" }}>{c.failover_to}</span>
-                  </span>
-                )}
-              </div>
-            )}
-          </div>
-        );
-      })}
     </div>
   );
 }
